chore(activitystream): refactor getUserTimeline
