+++ packages/server/modules/activitystream/services/index.js
@@ -107,39 +107,5 @@ module.exports = {
       items: results,
       cursor: results.length > 0 ? results[results.length - 1].time.toISOString() : null
     }
-  },
-
-  async getUserTimeline({ userId, after, before, cursor, limit }) {
-    if (!limit) {
-      limit = 200
-    }
-
-    let sqlFilters = ''
-    const sqlVariables = []
-    if (after) {
-      sqlFilters += ' AND time > ?'
-      sqlVariables.push(after)
-    }
-    if (before || cursor) {
-      sqlFilters += ' AND time < ?'
-      sqlVariables.push(before || cursor)
-    }
-
-    const dbRawQuery = `
-      SELECT act.*
-      FROM stream_acl acl
-      INNER JOIN stream_activity act ON acl."resourceId" = act."streamId"
-      WHERE acl."userId" = ? ${sqlFilters}
-      ORDER BY time DESC
-      LIMIT ?
-    `
-
-    sqlVariables.unshift(userId)
-    sqlVariables.push(limit)
-    const results = (await knex.raw(dbRawQuery, sqlVariables)).rows
-    return {
-      items: results,
-      cursor: results.length > 0 ? results[results.length - 1].time.toISOString() : null
-    }
   }
 }
